add leaflet option
