+++ package.json
@@ -14,12 +14,14 @@
     "axios": "^1.2.1",
     "bootstrap": "^5.2.3",
     "darkreader": "^4.9.58",
+    "leaflet": "^1.9.3",
     "plotly.js": "^2.17.0",
     "react": "^18.2.0",
     "react-bootstrap": "^2.6.0",
     "react-dom": "^18.2.0",
     "react-grid-layout": "^1.3.4",
     "react-icons": "^4.7.1",
+    "react-leaflet": "^4.2.0",
     "react-plotly.js": "^2.6.0",
     "react-pro-sidebar": "^1.0.0-alpha.9",
     "react-resize-detector": "^7.1.2",
@@ -58,6 +60,7 @@
     ]
   },
   "devDependencies": {
+    "@types/leaflet": "^1.9.0",
     "@types/react-grid-layout": "^1.3.2",
     "@types/react-plotly.js": "^2.5.2",
     "@types/styled-components": "^5.1.26",
