+++ src/pages/Maps.tsx
@@ -1,11 +1,8 @@
-
 import Map from "../components/tiles/Map";
-
-
 
 const Maps = () => {
     return (
-        <Map></Map>
+        <Map />
     );
 }
